requirements.txt
numpy
faiss-cpu
sentence-transformers
flask
openai
